+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add custom widgets: group::HGrid and group::VGrid.
 - Add lib search paths for Mageia 7 by @nsalguero.
 - Fix docs on special characters. 
+- Add feature: no-pango if rtl and cjk unicode support is not needed on linux/bsd.s
 
 ## [0.10.9] - 2020-11-15
 ### Changes